requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aioredis==2.0.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.2
pytz==2022.6
RPi.GPIO==0.7.1
typing_extensions==4.4.0
yarl==1.8.1
